Fixed a typo in Reservoir
